Removed MetadataProvider and instead metadata is being stamped on TokenProgramData
