+++ explorer/src/metaplex/ids.ts
@@ -45,10 +45,6 @@ export interface PublicKeyStringAndAccount<T> {
     account: AccountInfo<T>;
 }
 
-export const TOKEN_PROGRAM_ID = new PublicKey(
-    'TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA',
-);
-
 export const SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID = new PublicKey(
     'ATokenGPvbdGVxr1b2hvZbsiqW5xWH25efTNsLJA8knL',
 );
